Removed + from keywords.operators.symbols.actionscript so that an operator is now only a single character.
The reason for this is that since both / and * are in the group of characters matched, /** would match as a single operator, and since this is a longer match than /*, it would win over a comment start marker.

My change should not affect an actual sequence of operators, other than these are now matched as several individual operators, but in practice it will render the same.

Since I do not know AS, please feel free to educate me if the + had significant meaning.

git-svn-id: http://svn.textmate.org/trunk/ActionScript.tmbundle@350 dfb7d73b-c2ec-0310-8fea-fb051d288c6d